+++ apx_sdk/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Removed restriction on maximum version of `reqwest` dependency.
+- Changed error message in case of invalid `verificationMethod`.
 
 ### Fixed